fix conflict nav_graph
