+++ app/src/main/res/navigation/nav_graph.xml
@@ -173,4 +173,9 @@
         android:name="com.goldenowl.ecommerceapp.ui.Order.OrderDetailFragment"
         android:label="fragment_order_detail"
         tools:layout="@layout/fragment_order_detail" />
+    <fragment
+        android:id="@+id/searchFragment"
+        android:name="com.goldenowl.ecommerceapp.SearchFragment"
+        android:label="fragment_search"
+        tools:layout="@layout/fragment_search" />
 </navigation>
